[PROTO] update feature list, enable 64bit features
add new features for no filtering

fix feature size (uint32->uint64) in
sending config and feature_code_str

+++ daemon/daemon.c
@@ -714,8 +714,8 @@ static int targetServerHandler(int efd)
 
 		/* send config message to target process */
 		log.type = MSG_OPTION;
-		log.length = sprintf(log.data, "%lu",
-				     (unsigned long int)prof_session.conf.use_features0);
+		log.length = sprintf(log.data, "%llu",
+				     prof_session.conf.use_features0);
 		if (0 > send(manager.target[index].socket, &log,
 			     sizeof(log.type) + sizeof(log.length) + log.length,
 			     MSG_NOSIGNAL))
